server/requirements.txt
Flask==2.3.3
flask-sqlalchemy==3.0.3
flask_cors==1.8.0
PyMySQL==1.1.0
cryptography==41.0.4
python-dotenv==1.0.0
PyJWT==1.7.1